refactor(compass): move metric-series readers to the metrics module
Address PR review:
- series / latest / value_n_periods_ago / sum_last_n_periods now live as free
  functions in polar.metrics.aggregation, so any metrics consumer can reuse
  them and they're compartmentalised with Metrics rather than Compass. Kept as
  free functions (not methods on MetricsResponse) so the response schema stays
  data-only.
- MetricSignal.drivers_dimension is typed with a DriverDimension enum instead
  of a bare str.
