+++ server/polar/compass/signals.py
@@ -7,11 +7,19 @@ we compute them by reading the existing `metrics` service (which already merges
 Postgres + Tinybird, and handles auth, filtering and caching) over a window, then
 diffing periods in Python. Detectors depend only on `MetricSignal`, so swapping in
 dedicated pipes later is a signals-layer change, not a detector change.
+
+The per-period readers (`series`, `latest`, …) live on `MetricsResponse` in the
+metrics module, so they're reusable by any metrics consumer, not just Compass.
 """
 
 from dataclasses import dataclass, field
+from enum import StrEnum
 
-from polar.metrics.schemas import MetricsResponse
+
+class DriverDimension(StrEnum):
+    """What a driver breakdown is grouped by (e.g. per product)."""
+
+    product = "product"
 
 
 @dataclass(frozen=True)
@@ -23,7 +31,7 @@ class MetricSignal:
     baseline: float
     sample_n: int
     """Population behind the change (e.g. active subscriptions), for confidence."""
-    drivers_dimension: str | None = None
+    drivers_dimension: DriverDimension | None = None
     drivers: list[tuple[str, float]] = field(default_factory=list)
     """(label, contribution) pairs; empty until a breakdown pipe is wired in."""
 
@@ -36,30 +44,6 @@ class MetricSignal:
         if self.baseline == 0:
             return None
         return (self.current - self.baseline) / abs(self.baseline)
-
-
-def series(response: MetricsResponse, slug: str) -> list[float]:
-    """Ordered per-period values for a metric slug (periods are already sorted)."""
-    return [getattr(p, slug, None) or 0 for p in response.periods]
-
-
-def latest(response: MetricsResponse, slug: str) -> float:
-    values = series(response, slug)
-    return values[-1] if values else 0
-
-
-def value_n_periods_ago(response: MetricsResponse, slug: str, n: int) -> float | None:
-    """The value `n` periods before the latest, or None if the window is too short."""
-    values = series(response, slug)
-    index = len(values) - 1 - n
-    if index < 0:
-        return None
-    return values[index]
-
-
-def sum_last_n_periods(response: MetricsResponse, slug: str, n: int) -> float:
-    """Total of a metric over the most recent `n` periods (a rolling window)."""
-    return sum(series(response, slug)[-n:])
 
 
 def format_currency(cents: float) -> str:
